Finalizacion de endpoints

+++ backend/src/main/java/com/client/service_client/service/HotelService.java
@@ -3,7 +3,6 @@ package com.client.service_client.service;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.client.service_client.model.Hotel;
 import com.client.service_client.repository.HotelRepository;
@@ -20,13 +19,16 @@ public class HotelService {
         return hotelRepository.findAll();
     }
 
-    @Transactional
     public void save (Hotel hotel) {
         hotelRepository.save(hotel);
     }
 
-    @Transactional
     public void deleteById (String id) {
-        hotelRepository.deleteById(id);
+        if(hotelRepository.existsHotel(id)) {
+            hotelRepository.deleteById(id);
+        }
+        else {
+            throw new RuntimeException("Hotel not found");
+        }
     }
 }
